+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Bug fixes
 
+- [#86][GL-86] - Players may respawn at the wrong location after being revived
 - [#84][GL-84] - Cherry trees are not supported in the Timber and Fast Leaves Decay scenarios
 - [#83][GL-83] - The explosion-power option in the Double Dates scenario does not support floating-point values
 - [#79][GL-79] - Offline zombie is not always removed
 - [#78][GL-78] - Player inventory items are spawned at the wrong location when using the revive command
 
+[GL-86]: https://gitlab.com/uhccore/uhccore/-/issues/86
 [GL-84]: https://gitlab.com/uhccore/uhccore/-/issues/84
 [GL-83]: https://gitlab.com/uhccore/uhccore/-/issues/83
 [GL-79]: https://gitlab.com/uhccore/uhccore/-/issues/79